(feat) unify test files
